Movement of the INTEGRATION_TEST_SERVER to a better place.
