+++ linera-service/src/cli_wrappers/local_net.rs
@@ -29,6 +29,95 @@ use tonic_health::pb::{
 };
 use tracing::{info, warn};
 
+use linera_base::sync::Lazy;
+use async_lock::RwLock;
+use std::ops::Deref;
+use linera_storage_service::child::get_free_port;
+use linera_storage_service::client::service_config_from_endpoint;
+use linera_storage_service::common::ServiceStoreConfig;
+
+#[cfg(feature = "rocksdb")]
+use {
+    linera_views::rocks_db::{create_rocks_db_test_config, RocksDbStoreConfig},
+};
+
+
+struct IntegrationTestServerInternal {
+    pub service_config: ServiceStoreConfig,
+    _service_guard: StorageServiceGuard,
+    #[cfg(feature = "rocksdb")]
+    pub rocks_db_config: RocksDbStoreConfig,
+    #[cfg(feature = "rocksdb")]
+    _temp_dir: TempDir,
+}
+
+pub struct IntegrationTestServerConfig {
+    pub service_config: ServiceStoreConfig,
+    #[cfg(feature = "rocksdb")]
+    pub rocks_db_config: RocksDbStoreConfig,
+}
+
+impl IntegrationTestServerInternal {
+    async fn service_info() -> Result<(ServiceStoreConfig, StorageServiceGuard)> {
+        let endpoint = get_free_port().await.unwrap();
+        let service_config = service_config_from_endpoint(&endpoint)?;
+        let binary = get_service_storage_binary().await?.display().to_string();
+        let spanner = StorageServiceBuilder::new(&endpoint, binary);
+        let service_guard = spanner.run_service().await?;
+        Ok((service_config, service_guard))
+    }
+
+    #[cfg(feature = "rocksdb")]
+    pub async fn new() -> Result<Self> {
+        let (service_config, _service_guard) = Self::service_info().await?;
+        let (rocks_db_config, _temp_dir) = create_rocks_db_test_config().await;
+        Ok(Self { service_config, _service_guard, rocks_db_config, _temp_dir })
+    }
+
+    #[cfg(not(feature = "rocksdb"))]
+    pub async fn new() -> Result<Self> {
+        let (service_config, _service_guard) = Self::service_info().await?;
+        Ok(Self { service_config, _service_guard })
+    }
+
+    #[cfg(feature = "rocksdb")]
+    pub fn get_config(&self) -> IntegrationTestServerConfig {
+        let service_config = self.service_config.clone();
+        let rocks_db_config = self.rocks_db_config.clone();
+        IntegrationTestServerConfig { service_config, rocks_db_config }
+    }
+
+    #[cfg(not(feature = "rocksdb"))]
+    pub fn get_config(&self) -> IntegrationTestServerConfig {
+        let service_config = self.service_config.clone();
+        IntegrationTestServerConfig { service_config }
+    }
+}
+
+pub struct IntegrationTestServer {
+    internal_server: RwLock<Option<IntegrationTestServerInternal>>,
+}
+
+impl IntegrationTestServer {
+    pub fn new() -> Self {
+        Self { internal_server: RwLock::new(None) }
+    }
+
+    pub async fn get_config(&self) -> IntegrationTestServerConfig {
+        let mut w = self.internal_server.write().await;
+        if w.is_none() {
+            *w = Some(IntegrationTestServerInternal::new().await.expect("server"));
+        }
+        let Some(internal_server) = w.deref() else {
+            unreachable!();
+        };
+        internal_server.get_config()
+    }
+}
+
+// A static data to store the integration test server
+pub static INTEGRATION_TEST_SERVER: Lazy<IntegrationTestServer> = Lazy::new(|| IntegrationTestServer::new());
+
 /// The endpoint used for all the tests
 const END_TO_END_STORAGE_SERVICE_ENDPOINT: &str = "127.0.0.1:8742";
 
@@ -153,6 +242,7 @@ impl LineraNetConfig for LocalNetConfig {
     type Net = LocalNet;
 
     async fn instantiate(self) -> Result<(Self::Net, ClientWrapper)> {
+        let _server_config = INTEGRATION_TEST_SERVER.get_config().await;
         ensure!(
             self.num_shards == 1 || self.database != Database::RocksDb,
             "Multiple shards not supported with RocksDB"
@@ -389,27 +479,33 @@ impl LocalNet {
     }
 
     async fn run_server(&mut self, validator: usize, shard: usize) -> Result<Child> {
+        let namespace = match self.database {
+            Database::Service => format!("{}_server_{}_db", self.table_name, validator),
+            Database::RocksDb => format!("server_{}_{}.db", validator, shard),
+            Database::DynamoDb => format!("{}_server_{}.db", self.table_name, validator),
+            Database::ScyllaDb => format!("{}_server_{}_db", self.table_name, validator),
+        };
         let (storage, key) = match self.database {
             Database::Service => (
                 format!(
-                    "service:http://{}:{}_server_{}_db",
-                    END_TO_END_STORAGE_SERVICE_ENDPOINT, self.table_name, validator
+                    "service:http://{}:{}",
+                    END_TO_END_STORAGE_SERVICE_ENDPOINT, namespace,
                 ),
                 (validator, 0),
             ),
             Database::RocksDb => (
-                format!("rocksdb:server_{}_{}.db", validator, shard),
+                format!("rocksdb:{}", namespace),
                 (validator, shard),
             ),
             Database::DynamoDb => (
                 format!(
-                    "dynamodb:{}_server_{}.db:localstack",
-                    self.table_name, validator
+                    "dynamodb:{}:localstack",
+                    namespace,
                 ),
                 (validator, 0),
             ),
             Database::ScyllaDb => (
-                format!("scylladb:{}_server_{}_db", self.table_name, validator),
+                format!("scylladb:{}", namespace),
                 (validator, 0),
             ),
         };
